Make multiple test functions if only write methods are given.
