+++ rest_tester/function/read.py
@@ -1,29 +1,16 @@
-import copy
-
 import requests
 
-from rest_tester.setting import ParameterSet, TestMethod, UnsupportedMethodError
-from . import TestFunctionBuilder, TestFunction
+from rest_tester.setting import TestMethod, UnsupportedMethodError
+from . import TestFunctionBuilder
 
 
 class ReadTestFunctionBuilder(TestFunctionBuilder):
     """Build function that checks the response from READ (e.g., GET) method."""
     def build(self):
-        request = self._setting.read_request
-        response = self._setting.read_response
+        read_request = self._setting.read_request
+        read_response = self._setting.read_response
 
-        test_function_list = []
-
-        param_set_list = ParameterSet.generate(request.params)
-        for param_set in param_set_list:
-            curr_request = copy.deepcopy(request)
-            curr_request.params = param_set
-
-            test_function_name = self._generate_name(self._name_prefix, curr_request)
-            test_function = self._build_test_function(curr_request, response)
-            test_function_list.append(TestFunction(test_function_name, test_function))
-
-        return test_function_list
+        return self._build_test_function_list(read_request, read_response)
 
     def _get_actual_response(self, request, params):
         read_method = self._setting.method.read_method
