+++ rest_tester/main.py
@@ -7,7 +7,6 @@ import unittest
 from rest_tester.function import TestFunctionBuilderFactory
 from rest_tester.function.fail import FailTestFunctionBuilder
 from rest_tester.setting import TestSetting
-from rest_tester.utils import convert_to_list
 
 
 class TestsContainer(unittest.TestCase):
@@ -57,8 +56,8 @@ def main(argv):
             continue
 
         test_function_builder = TestFunctionBuilderFactory.get_builder(setting, file_name)
-        test_functions = test_function_builder.build()
-        for test_function in convert_to_list(test_functions):
+        test_function_list = test_function_builder.build()
+        for test_function in test_function_list:
             add_function_to_container(test_function)
 
     suite = unittest.makeSuite(TestsContainer)
