+++ test/function/resources/test_function_write_put_only_multi_params.json
@@ -1,0 +1,19 @@
+{
+  "put": {
+    "request": {
+      "url": "http://json-server:3000/posts/1",
+      "params": {
+        "id": [3, 4, 5]
+      },
+      "data": {
+        "title": "foo",
+        "body": "bar",
+        "userId": 1
+      },
+      "timeout" : 10
+    },
+    "response": {
+      "statusCode": 200
+    }
+  }
+}
