+++ test/function/test_read.py
@@ -1,12 +1,12 @@
 import os
 import unittest
-import time
 
 from jsonschema import ValidationError
 
 from rest_tester.function.read import ReadTestFunctionBuilder
 from rest_tester.setting import TestSetting
 from test.helper import load_json_data
+from . import run_test_function_list
 
 
 class TestReadTestFunctionBuilder(unittest.TestCase):
@@ -16,14 +16,14 @@ class TestReadTestFunctionBuilder(unittest.TestCase):
         json_file = '%s/resources/test_function_read_get.json' % self.current_dir_path
         test_function_list = self.generate_test_functions(json_file)
 
-        self.run_test_function_list(test_function_list, self)
+        run_test_function_list(test_function_list, self)
 
     def test_build_unexpected_status_code(self):
         json_file = '%s/resources/test_function_read_get_unexpected_status_code.json' % self.current_dir_path
         test_function_list = self.generate_test_functions(json_file)
 
         try:
-            self.run_test_function_list(test_function_list, self)
+            run_test_function_list(test_function_list, self)
         except AssertionError:
             pass
         else:
@@ -34,7 +34,7 @@ class TestReadTestFunctionBuilder(unittest.TestCase):
         test_function_list = self.generate_test_functions(json_file)
 
         try:
-            self.run_test_function_list(test_function_list, self)
+            run_test_function_list(test_function_list, self)
         except ValidationError:
             pass
         else:
@@ -43,19 +43,12 @@ class TestReadTestFunctionBuilder(unittest.TestCase):
     @staticmethod
     def generate_test_functions(json_file):
         json_data = load_json_data(json_file)
-        setting = TestSetting(json_data)
 
+        setting = TestSetting(json_data)
         file_name = os.path.basename(json_file)
 
         builder = ReadTestFunctionBuilder(setting, file_name)
         return builder.build()
-
-    @staticmethod
-    def run_test_function_list(test_function_list, test_self):
-        time.sleep(1)
-
-        for test_function in test_function_list:
-            test_function.test_function(test_self)
 
 if __name__ == '__main__':
     unittest.main()
