+++ test/function/test_write.py
@@ -1,12 +1,12 @@
 import os
 import unittest
-import time
 
 from jsonschema import ValidationError
 
 from rest_tester.function.write import WriteTestFunctionBuilder
 from rest_tester.setting import TestSetting
 from test.helper import load_json_data
+from . import run_test_function_list
 
 
 class TestWriteTestFunctionBuilder(unittest.TestCase):
@@ -14,40 +14,45 @@ class TestWriteTestFunctionBuilder(unittest.TestCase):
 
     def test_build_put(self):
         json_file = '%s/resources/test_function_write_put.json' % self.current_dir_path
-        test_function = self.generate_test_function(json_file)
-        self.run_test_function(test_function, self)
+        test_function_list = self.generate_test_function(json_file)
+        run_test_function_list(test_function_list, self)
 
     def test_build_post(self):
         json_file = '%s/resources/test_function_write_post.json' % self.current_dir_path
-        test_function = self.generate_test_function(json_file)
-        self.run_test_function(test_function, self)
+        test_function_list = self.generate_test_function(json_file)
+        run_test_function_list(test_function_list, self)
 
     def test_build_put_only(self):
         json_file = '%s/resources/test_function_write_put_only.json' % self.current_dir_path
-        test_function = self.generate_test_function(json_file)
-        self.run_test_function(test_function, self)
+        test_function_list = self.generate_test_function(json_file)
+        run_test_function_list(test_function_list, self)
+
+    def test_build_put_only_multi_params(self):
+        json_file = '%s/resources/test_function_write_put_only_multi_params.json' % self.current_dir_path
+        test_function_list = self.generate_test_function(json_file)
+        run_test_function_list(test_function_list, self)
 
     def test_build_post_only(self):
         json_file = '%s/resources/test_function_write_post.json' % self.current_dir_path
-        test_function = self.generate_test_function(json_file)
-        self.run_test_function(test_function, self)
+        test_function_list = self.generate_test_function(json_file)
+        run_test_function_list(test_function_list, self)
 
     def test_build_patch_only(self):
         json_file = '%s/resources/test_function_write_patch_only.json' % self.current_dir_path
-        test_function = self.generate_test_function(json_file)
-        self.run_test_function(test_function, self)
+        test_function_list = self.generate_test_function(json_file)
+        run_test_function_list(test_function_list, self)
 
     def test_build_delete_only(self):
         json_file = '%s/resources/test_function_write_delete_only.json' % self.current_dir_path
-        test_function = self.generate_test_function(json_file)
-        self.run_test_function(test_function, self)
+        test_function_list = self.generate_test_function(json_file)
+        run_test_function_list(test_function_list, self)
 
     def test_build_unexpected_status_code(self):
         json_file = '%s/resources/test_function_write_unexpected_status_code.json' % self.current_dir_path
-        test_function = self.generate_test_function(json_file)
+        test_function_list = self.generate_test_function(json_file)
 
         try:
-            self.run_test_function(test_function, self)
+            run_test_function_list(test_function_list, self)
         except AssertionError:
             pass
         else:
@@ -55,10 +60,10 @@ class TestWriteTestFunctionBuilder(unittest.TestCase):
 
     def test_build_unexpected_get(self):
         json_file = '%s/resources/test_function_write_unexpected_get.json' % self.current_dir_path
-        test_function = self.generate_test_function(json_file)
+        test_function_list = self.generate_test_function(json_file)
 
         try:
-            self.run_test_function(test_function, self)
+            run_test_function_list(test_function_list, self)
         except ValidationError:
             pass
         else:
@@ -73,12 +78,6 @@ class TestWriteTestFunctionBuilder(unittest.TestCase):
 
         builder = WriteTestFunctionBuilder(setting, file_name)
         return builder.build()
-
-    @staticmethod
-    def run_test_function(test_function, test_self):
-        time.sleep(1)
-
-        test_function.test_function(test_self)
 
 if __name__ == '__main__':
     unittest.main()
